+++ backend/is24/src/main/java/com/yin/bigdata/is24/REExpose.java
@@ -9,7 +9,7 @@ public class REExpose {
 	private double latitude ;
 	private double size ;
 	private double rooms ;
-	private double constrYear ;
+	private int constrYear ;
 	private String condition;
 	private String heatingType;
 	private boolean balcony;
@@ -44,10 +44,10 @@ public class REExpose {
 	public void setRooms(double rooms) {
 		this.rooms = rooms;
 	}
-	public double getConstrYear() {
+	public int getConstrYear() {
 		return constrYear;
 	}
-	public void setConstrYear(double constrYear) {
+	public void setConstrYear(int constrYear) {
 		this.constrYear = constrYear;
 	}
 	public String getCondition() {
